requirements.txt
anthropic>=0.8.0
python-dotenv>=1.0.0
pyupbit>=0.2.33
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
matplotlib>=3.7.0
websocket-client>=1.6.0
schedule>=1.2.0
pyyaml>=6.0